refactor: update SubmissionCard import paths and create new SubmissionCard component
