+++ apps/nova/src/components/common/SubmissionCard.tsx
